added react hook forms, removed layout and fixed array.ts for Login Screen
