+++ cookbook-react/src/app/contexts/UserContext/reducer.ts
@@ -10,7 +10,7 @@ export interface UserState {
 }
 
 export interface Credentials {
-  username: string;
+  email: string;
   password: string;
 }
